[refactor] テストファイルをクラスで分割
